Safari-611.1.10/LayoutTests/svg/dom/set-currentScale-nonfinite-expected.txt

This test checks that setting currentScale fails for non-finite values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgvar.currentScale is 1.0
PASS Setting currentScale to undefined threw an exception.
PASS svgvar.currentScale is 1.0
PASS Setting currentScale to NaN threw an exception.
PASS svgvar.currentScale is 1.0
PASS Setting currentScale to Infinity threw an exception.
PASS svgvar.currentScale is 1.0
PASS Setting currentScale to a string threw an exception.
PASS svgvar.currentScale is 1.0
PASS successfullyParsed is true

TEST COMPLETE
